Added Enemy Prototype
Started working on the prefab that will become each enemy. Since the visual representation has not been created yet, a white circle will represent the enemy. The enemy can now move around the warehouse, and as soon as the player character is implemented, the behaviour/component will be updated to check if the enemy has collided with the player. As of now the enemy moves along a set path, just like our plan outlined.